requirements.txt
torch>=2.0.0
torchvision>=0.15.0
monai>=1.3.0
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
gudhi>=3.7.1
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
SimpleITK>=2.2.0
nibabel>=5.1.0
scikit-learn>=1.3.0